separate creation of the db connection and test content
this way the workflow engine can be initialized before any content is created, in particular the root node.  with this the workaround to get a "clean" root (from f12f9d9d) or rather its update (from c494bca1) can be removed again.  this is because the only difference from the root node created in the `connection` fixture before and the "clean" one was the workflow state, in fact.

also note that a number of workflow-related tests are currently still failing...